requirements.txt
# Production dependencies
requests==2.31.0
python-dotenv==1.0.0
cdp-sdk>=0.11.0
cryptography>=46.0.0
coinbase-advanced-py>=1.8.0

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-html==4.1.1

# Code quality tools
pylint==3.0.3
flake8==6.1.0
black==23.12.1

# Security tools
safety==2.3.5
bandit==1.7.5
